+++ boa_engine/src/builtins/typed_array/integer_indexed_object.rs
@@ -32,18 +32,18 @@ unsafe impl Trace for ContentType {
 pub struct IntegerIndexed {
     viewed_array_buffer: Option<JsObject>,
     typed_array_name: TypedArrayKind,
-    byte_offset: usize,
-    byte_length: usize,
-    array_length: usize,
+    byte_offset: u64,
+    byte_length: u64,
+    array_length: u64,
 }
 
 impl IntegerIndexed {
     pub(crate) fn new(
         viewed_array_buffer: Option<JsObject>,
         typed_array_name: TypedArrayKind,
-        byte_offset: usize,
-        byte_length: usize,
-        array_length: usize,
+        byte_offset: u64,
+        byte_length: u64,
+        array_length: u64,
     ) -> Self {
         Self {
             viewed_array_buffer,
@@ -105,12 +105,12 @@ impl IntegerIndexed {
     }
 
     /// Get the integer indexed object's byte offset.
-    pub(crate) fn byte_offset(&self) -> usize {
+    pub(crate) fn byte_offset(&self) -> u64 {
         self.byte_offset
     }
 
     /// Set the integer indexed object's byte offset.
-    pub(crate) fn set_byte_offset(&mut self, byte_offset: usize) {
+    pub(crate) fn set_byte_offset(&mut self, byte_offset: u64) {
         self.byte_offset = byte_offset;
     }
 
@@ -130,22 +130,22 @@ impl IntegerIndexed {
     }
 
     /// Get the integer indexed object's byte length.
-    pub fn byte_length(&self) -> usize {
+    pub fn byte_length(&self) -> u64 {
         self.byte_length
     }
 
     /// Set the integer indexed object's byte length.
-    pub(crate) fn set_byte_length(&mut self, byte_length: usize) {
+    pub(crate) fn set_byte_length(&mut self, byte_length: u64) {
         self.byte_length = byte_length;
     }
 
     /// Get the integer indexed object's array length.
-    pub fn array_length(&self) -> usize {
+    pub fn array_length(&self) -> u64 {
         self.array_length
     }
 
     /// Set the integer indexed object's array length.
-    pub(crate) fn set_array_length(&mut self, array_length: usize) {
+    pub(crate) fn set_array_length(&mut self, array_length: u64) {
         self.array_length = array_length;
     }
 }
